♻️ refactor: preserve search mappings in versioned migration batches
