+++ packages/database/src/repositories/ftsSearchDocument/__tests__/schemaSnapshots.ts
@@ -1,0 +1,72 @@
+export interface FtsSearchIndexSchemaSnapshot {
+  fingerprint: string;
+  schemaVersion: number;
+}
+
+/**
+ * Approved physical index schemas, keyed by the source migration batch that introduced them.
+ * Published entries are immutable; append a batch when a mapping or shared analysis changes.
+ */
+export const FTS_SEARCH_INDEX_SCHEMA_SNAPSHOTS: Record<
+  string,
+  Record<string, FtsSearchIndexSchemaSnapshot>
+> = {
+  '0000-initial-search-indexes': {
+    agents: {
+      fingerprint: '6fd440be1df5ee1437f4a4ec7d22556d1c855ce43bd7fda61ecf0dae32ede233',
+      schemaVersion: 1,
+    },
+    chatGroups: {
+      fingerprint: '5b086a9c74ecf70d27334b12cae2de5c5b6412bdc4ad79290514b723b2ec23e2',
+      schemaVersion: 1,
+    },
+    documents: {
+      fingerprint: '6d564684eb449d0a70b3c354fa6d9fefe4948cee09d8b46335a7069ac064cde8',
+      schemaVersion: 1,
+    },
+    files: {
+      fingerprint: '709e38c13c4684ffc4cb5aec00b4cd276b61a23f0387595bae02466a30291ffc',
+      schemaVersion: 1,
+    },
+    knowledgeBases: {
+      fingerprint: 'db6e7e1027953600ca9d5d1147f6138ca6868fcd9c58e30e1ce75d2813027320',
+      schemaVersion: 1,
+    },
+    memoryActivities: {
+      fingerprint: 'ea4665bd0026a75719da8580a350245a8deafa85591d36b577e264e84e32d526',
+      schemaVersion: 1,
+    },
+    memoryContexts: {
+      fingerprint: '1b9c24a3ab665a84686c766568d976193f670d469fbd29ff30ab2736f3d23b84',
+      schemaVersion: 1,
+    },
+    memoryExperiences: {
+      fingerprint: '9748bd7a25f0f53162ad90e3742244e689b4023c863ad71c226b7d0a9763444c',
+      schemaVersion: 1,
+    },
+    memoryIdentities: {
+      fingerprint: '491c3b1a096644552e1486cf6862899bd57042df787f01319442d9e615da0c20',
+      schemaVersion: 1,
+    },
+    memoryPreferences: {
+      fingerprint: 'e62b555d4548b7f880a6700e0662a3b832a3aa052ad518c869881e1ce99660f7',
+      schemaVersion: 1,
+    },
+    messages: {
+      fingerprint: 'c1e0a4ea89bd0c3ad7776287deb05eb82a8d286dc851d112a60c729839bc0d4a',
+      schemaVersion: 1,
+    },
+    personaDocuments: {
+      fingerprint: '0e627cf7a6a0ce38ae8b06affc5ad6b30c3c443bf93ae19df743f2b652ec323b',
+      schemaVersion: 1,
+    },
+    topics: {
+      fingerprint: '362f4b72197fefee4ffb49298aa134c6da6fba9885e7a23840eb55f053a689b8',
+      schemaVersion: 1,
+    },
+    userMemories: {
+      fingerprint: 'df04d074c1f56b5934891b8612b804917c09302a9042015c394216513093d6e9',
+      schemaVersion: 1,
+    },
+  },
+};
