+++ packages/database/src/repositories/ftsSearchDocument/mappings.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
+import { FTS_SEARCH_INDEX_SCHEMA_SNAPSHOTS } from './__tests__/schemaSnapshots';
 import { getFtsSearchIndexSchemaFingerprint } from './fingerprint';
 import {
   FTS_SEARCH_INDEX_ANALYSIS,
@@ -9,78 +10,12 @@ import {
   getFtsSearchIndexSchemaVersion,
   getFtsSearchPhysicalIndexName,
 } from './mappings';
+import { FTS_SEARCH_CURRENT_MAPPING_MIGRATIONS } from './migration';
 import {
   FTS_SEARCH_DOCUMENT_ENTITIES,
   FTS_SEARCH_DOCUMENT_SCHEMAS,
   FTS_SEARCH_MEMORY_DOCUMENT_ENTITIES,
 } from './zodSchema';
-
-/**
- * Declared generation of every entity index. When a mapping or the shared analysis changes, its
- * fingerprint changes too; the gate below then fails until `schemaVersion` is bumped and this
- * snapshot is updated in the same change, so a live index can never silently drift from the code.
- */
-const FTS_SEARCH_INDEX_SCHEMA_SNAPSHOT: Record<
-  (typeof FTS_SEARCH_DOCUMENT_ENTITIES)[number],
-  { fingerprint: string; schemaVersion: number }
-> = {
-  agents: {
-    fingerprint: '6fd440be1df5ee1437f4a4ec7d22556d1c855ce43bd7fda61ecf0dae32ede233',
-    schemaVersion: 1,
-  },
-  chatGroups: {
-    fingerprint: '5b086a9c74ecf70d27334b12cae2de5c5b6412bdc4ad79290514b723b2ec23e2',
-    schemaVersion: 1,
-  },
-  documents: {
-    fingerprint: '6d564684eb449d0a70b3c354fa6d9fefe4948cee09d8b46335a7069ac064cde8',
-    schemaVersion: 1,
-  },
-  files: {
-    fingerprint: '709e38c13c4684ffc4cb5aec00b4cd276b61a23f0387595bae02466a30291ffc',
-    schemaVersion: 1,
-  },
-  knowledgeBases: {
-    fingerprint: 'db6e7e1027953600ca9d5d1147f6138ca6868fcd9c58e30e1ce75d2813027320',
-    schemaVersion: 1,
-  },
-  memoryActivities: {
-    fingerprint: 'ea4665bd0026a75719da8580a350245a8deafa85591d36b577e264e84e32d526',
-    schemaVersion: 1,
-  },
-  memoryContexts: {
-    fingerprint: '1b9c24a3ab665a84686c766568d976193f670d469fbd29ff30ab2736f3d23b84',
-    schemaVersion: 1,
-  },
-  memoryExperiences: {
-    fingerprint: '9748bd7a25f0f53162ad90e3742244e689b4023c863ad71c226b7d0a9763444c',
-    schemaVersion: 1,
-  },
-  memoryIdentities: {
-    fingerprint: '491c3b1a096644552e1486cf6862899bd57042df787f01319442d9e615da0c20',
-    schemaVersion: 1,
-  },
-  memoryPreferences: {
-    fingerprint: 'e62b555d4548b7f880a6700e0662a3b832a3aa052ad518c869881e1ce99660f7',
-    schemaVersion: 1,
-  },
-  messages: {
-    fingerprint: 'c1e0a4ea89bd0c3ad7776287deb05eb82a8d286dc851d112a60c729839bc0d4a',
-    schemaVersion: 1,
-  },
-  personaDocuments: {
-    fingerprint: '0e627cf7a6a0ce38ae8b06affc5ad6b30c3c443bf93ae19df743f2b652ec323b',
-    schemaVersion: 1,
-  },
-  topics: {
-    fingerprint: '362f4b72197fefee4ffb49298aa134c6da6fba9885e7a23840eb55f053a689b8',
-    schemaVersion: 1,
-  },
-  userMemories: {
-    fingerprint: 'df04d074c1f56b5934891b8612b804917c09302a9042015c394216513093d6e9',
-    schemaVersion: 1,
-  },
-};
 
 describe('search index schema generations', () => {
   it.each(FTS_SEARCH_DOCUMENT_ENTITIES)(
@@ -90,27 +25,14 @@ describe('search index schema generations', () => {
         fingerprint: getFtsSearchIndexSchemaFingerprint(entity),
         schemaVersion: getFtsSearchIndexSchemaVersion(entity),
       };
-      const snapshot = FTS_SEARCH_INDEX_SCHEMA_SNAPSHOT[entity];
-
-      if (actual.schemaVersion === snapshot.schemaVersion) {
-        expect(
-          actual.fingerprint,
-          `The ${entity} mapping or shared analysis changed without bumping schemaVersion. Elasticsearch cannot apply this change to the live v${snapshot.schemaVersion} index; bump schemaVersion and refresh FTS_SEARCH_INDEX_SCHEMA_SNAPSHOT in the same change.`,
-        ).toBe(snapshot.fingerprint);
-        return;
-      }
+      const migrationId = FTS_SEARCH_CURRENT_MAPPING_MIGRATIONS[entity].id;
+      const snapshot = FTS_SEARCH_INDEX_SCHEMA_SNAPSHOTS[migrationId]?.[entity];
 
       expect(
-        actual.schemaVersion,
-        `${entity} schemaVersion must only move forward by one generation per change.`,
-      ).toBe(snapshot.schemaVersion + 1);
-      expect(
-        actual.fingerprint,
-        `${entity} schemaVersion was bumped but the mapping fingerprint is unchanged; a new generation must carry a real mapping or analysis change.`,
-      ).not.toBe(snapshot.fingerprint);
-      expect.fail(
-        `${entity} moved to schemaVersion ${actual.schemaVersion}; refresh FTS_SEARCH_INDEX_SCHEMA_SNAPSHOT with fingerprint ${actual.fingerprint}.`,
-      );
+        snapshot,
+        `Add the approved ${migrationId}/${entity} schemaVersion and fingerprint to FTS_SEARCH_INDEX_SCHEMA_SNAPSHOTS.`,
+      ).toBeDefined();
+      expect(actual).toEqual(snapshot);
     },
   );
