+++ packages/database/src/repositories/ftsSearchDocument/migration/0000-initial-search-indexes/mapping.ts
@@ -1,0 +1,315 @@
+import type { FtsSearchMappingMigration } from '../types';
+import { analysis } from './analysis';
+import {
+  boolean,
+  date,
+  fileNameText,
+  icuText,
+  integer,
+  keyword,
+  memoryText,
+  memoryTextWithRaw,
+  mixedText,
+  ownershipProperties,
+  projectionMetadataProperties,
+  timestampProperties,
+} from './fields';
+
+/** Initial physical snapshots. Published batches retain their fields and analysis unchanged. */
+export const initialSearchIndexes = {
+  id: '0000-initial-search-indexes',
+  analysis,
+  definitions: {
+    agents: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...ownershipProperties,
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          description: mixedText,
+          id: keyword,
+          slug: icuText,
+          system_role: mixedText,
+          tags: icuText,
+          title: mixedText,
+          virtual: boolean,
+        },
+      },
+      schemaVersion: 1,
+    },
+    chatGroups: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...ownershipProperties,
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          content: mixedText,
+          description: mixedText,
+          group_id: keyword,
+          id: keyword,
+          title: mixedText,
+        },
+      },
+      schemaVersion: 1,
+    },
+    documents: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...ownershipProperties,
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          content: mixedText,
+          description: mixedText,
+          file_id: keyword,
+          file_type: keyword,
+          id: keyword,
+          knowledge_base_id: keyword,
+          knowledge_base_ids: keyword,
+          parent_id: keyword,
+          slug: icuText,
+          source_type: keyword,
+          title: mixedText,
+          total_char_count: integer,
+        },
+      },
+      schemaVersion: 1,
+    },
+    files: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...ownershipProperties,
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          file_type: keyword,
+          id: keyword,
+          knowledge_base_ids: keyword,
+          name: fileNameText,
+          size: integer,
+          source: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    knowledgeBases: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...ownershipProperties,
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          description: mixedText,
+          id: keyword,
+          is_public: boolean,
+          name: icuText,
+          type: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    memoryActivities: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          ends_at: date,
+          feedback: memoryText,
+          id: keyword,
+          narrative: memoryText,
+          notes: memoryText,
+          parent_details: memoryText,
+          parent_memory_categories: keyword,
+          parent_summary: memoryText,
+          parent_tags: keyword,
+          parent_title: memoryText,
+          starts_at: date,
+          status: keyword,
+          tags: keyword,
+          type: keyword,
+          user_id: keyword,
+          user_memory_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    memoryContexts: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          current_status: memoryTextWithRaw,
+          description: memoryText,
+          id: keyword,
+          parent_text: memoryText,
+          parent_memory_categories: keyword,
+          parent_tags: keyword,
+          tags: keyword,
+          title: memoryText,
+          type: keyword,
+          user_id: keyword,
+          user_memory_ids: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    memoryExperiences: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          action: memoryText,
+          captured_at: date,
+          id: keyword,
+          key_learning: memoryText,
+          parent_details: memoryText,
+          parent_memory_categories: keyword,
+          parent_summary: memoryText,
+          parent_tags: keyword,
+          parent_title: memoryText,
+          possible_outcome: memoryText,
+          reasoning: memoryText,
+          situation: memoryText,
+          tags: keyword,
+          type: keyword,
+          user_id: keyword,
+          user_memory_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    memoryIdentities: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          description: memoryText,
+          episodic_date: date,
+          id: keyword,
+          parent_details: memoryText,
+          parent_memory_categories: keyword,
+          parent_summary: memoryText,
+          parent_tags: keyword,
+          parent_title: memoryText,
+          relationship: keyword,
+          role: memoryText,
+          tags: keyword,
+          type: keyword,
+          user_id: keyword,
+          user_memory_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    memoryPreferences: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          conclusion_directives: memoryText,
+          id: keyword,
+          parent_details: memoryText,
+          parent_memory_categories: keyword,
+          parent_summary: memoryText,
+          parent_tags: keyword,
+          parent_title: memoryText,
+          suggestions: memoryText,
+          tags: keyword,
+          type: keyword,
+          user_id: keyword,
+          user_memory_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    messages: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          agent_id: keyword,
+          content: mixedText,
+          group_id: keyword,
+          id: keyword,
+          role: keyword,
+          session_id: keyword,
+          summary: mixedText,
+          thread_id: keyword,
+          topic_id: keyword,
+          user_id: keyword,
+          workspace_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    personaDocuments: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          id: keyword,
+          persona: memoryText,
+          profile: keyword,
+          tagline: memoryText,
+          user_id: keyword,
+          version: integer,
+        },
+      },
+      schemaVersion: 1,
+    },
+    topics: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          agent_id: keyword,
+          content: mixedText,
+          description: mixedText,
+          group_id: keyword,
+          id: keyword,
+          session_id: keyword,
+          status: keyword,
+          title: mixedText,
+          user_id: keyword,
+          workspace_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+    userMemories: {
+      mappings: {
+        dynamic: 'strict',
+        properties: {
+          ...projectionMetadataProperties,
+          ...timestampProperties,
+          captured_at: date,
+          details: memoryText,
+          id: keyword,
+          memory_category: keyword,
+          memory_layer: keyword,
+          status: keyword,
+          summary: memoryText,
+          tags: keyword,
+          title: memoryText,
+          user_id: keyword,
+        },
+      },
+      schemaVersion: 1,
+    },
+  },
+} as const satisfies FtsSearchMappingMigration;
